pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "swilcox", email = "steven@wilcoxzone.com"},
]
dependencies = [
    "django>=4.2.6",
    "django-ninja>=0.22.2",
    "dj-database-url>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "django-dramatiq>=0.11.5",
    "dramatiq[all]>=1.15.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

